pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "kriscv"
version = "0.1.6"
description = "K tooling for the RISC-V architecture"
authors = [
    "Runtime Verification, Inc. <contact@runtimeverification.com>",
]

[tool.poetry.plugins.kdist]
riscv-semantics = "kriscv.kdist.plugin"

[tool.poetry.dependencies]
python = "^3.10"
pyk = { git = "https://github.com/runtimeverification/k.git", tag = "v7.0.110", subdirectory = "pyk" }
pyyaml = "^6.0.1"
types-pyyaml = "^6.0.12.20240311"
filelock = "^3.14.0"

[tool.poetry.group.dev.dependencies]
autoflake = "*"
black = "*"
flake8 = "*"
flake8-bugbear = "*"
flake8-comprehensions = "*"
flake8-quotes = "*"
flake8-type-checking = "*"
isort = "*"
mypy = "*"
pep8-naming = "*"
pytest = "*"
pytest-cov = "*"
pytest-mock = "*"
pytest-xdist = "*"
pyupgrade = "*"

[tool.isort]
profile = "black"
line_length = 120

[tool.autoflake]
recursive = true
expand-star-imports = true
remove-all-unused-imports = true
ignore-init-module-imports = true
remove-duplicate-keys = true
remove-unused-variables = true

[tool.black]
line-length = 120
skip-string-normalization = true

[tool.mypy]
disallow_untyped_defs = true
